requirements.txt
python-fasthtml
requests
pydantic
openai
anthropic
markupsafe